updates setter

+++ State.java
@@ -12,7 +12,7 @@ public class State extends Country{
 		return LandArea;
 	}
 	
-	public void setArea(int a)
+	public void setArea(double a)
 	{
 		LandArea = a;
 	}
